Remove `musl`'s `vfork` functionality and use our own asm wrappers
We are now using our own `vfork` assembly wrappers in the unikernel
so drop current `vfork` assembly functions that we cannot use anyway
for native syscalls and instead copy/reuse those defined by
us in the core repository.

Do this copy-paste so that in case if int the future there will appear
differences between what we do in the core in those wrappers and what
we need for musl, the musl wrappers stay the same.

Signed-off-by: Sergiu Moga <sergiu@unikraft.io>
Approved-by: Michalis Pappas <michalis@unikraft.io>
Reviewed-by: Andrei Tatar <andrei@unikraft.io>
Reviewed-by: Michalis Pappas <michalis@unikraft.io>
GitHub-Closes: #86